+++ src/main/java/com/ss/touragency/dao/ICrudDao.java
@@ -5,14 +5,32 @@ import java.util.List;
 
 public interface ICrudDao<SomeEntity> {
 
+    /**
+     * Query for insert entity to database
+     *
+     * @param entity Object from package entity
+     */
     //create
     void insert(SomeEntity entity);
 
 
+    /**
+     * Query for select all entities from database
+     *
+     * @return List of entities
+     * @throws SQLException
+     */
     //read
     List<SomeEntity> selectAll() throws SQLException;
 
-    SomeEntity selectById(int id) throws SQLException;
+    /**
+     * Search entity by Id from database
+     *
+     * @param id Id entity in database
+     * @return some entity found by Id
+     * @throws SQLException
+     */
+    SomeEntity selectById(Long id) throws SQLException;
 
     // update
     void updateById(SomeEntity entity, Long id);
